requirements.txt
streamlit
requests
beautifulsoup4
python-docx
pillow
